Implement user authorization to blog deletion

+++ part4/controllers/blogs.js
@@ -49,6 +49,15 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const userId = decodedToken.id
+  if ( !( userId && blog.user.toString() === userId.toString() ) ) {
+    return response
+      .status(401)
+      .json({ error: 'missing or invalid token' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
